requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.3.0
